+++ cmd/template/cluster/provider/templates/capa/functions.go
@@ -34,7 +34,11 @@ func GenerateClusterValues(flagInputs ClusterConfig) (string, error) {
 		}
 	}
 
-	finalConfigString, err := yaml.Marshal(flagInputs)
+	if metadata, ok := flagConfigData["global"].(map[string]interface{})["metadata"].(map[string]interface{}); ok {
+		metadata["preventDeletion"] = flagInputs.Global.Metadata.PreventDeletion
+	}
+
+	finalConfigString, err := yaml.Marshal(flagConfigData)
 	if err != nil {
 		return "", microerror.Mask(err)
 	}