+++ cmd/template/cluster/provider/templates/capz/types.go
@@ -12,9 +12,10 @@ type Global struct {
 }
 
 type Metadata struct {
-	Name         string `json:"name,omitempty"`
-	Description  string `json:"description,omitempty"`
-	Organization string `json:"organization,omitempty"`
+	Name            string `json:"name,omitempty"`
+	Description     string `json:"description,omitempty"`
+	Organization    string `json:"organization,omitempty"`
+	PreventDeletion bool   `json:"preventDeletion,omitempty"`
 }
 
 type DefaultAppsConfig struct {